portfolio/backend/requirements.txt
flask
jinja2
python-dotenv
werkzeug
python-json-logger
flask-cors
